+++ server/src-lib/Hasura/Backends/BigQuery/DDL/Source.hs
@@ -173,7 +173,8 @@ restTableReferenceToTableName RestTableReference {..} =
 postDropSourceHook ::
   (MonadIO m) =>
   BigQuerySourceConfig ->
+  TableEventTriggers 'BigQuery ->
   m ()
-postDropSourceHook _ =
+postDropSourceHook _ _ =
   -- On BigQuery we don't keep connections open.
   pure ()