+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Bug fixes and improvements
 
+- server: fix bug where hasura SQL trigger was not dropped when MSSQL source is dropped
+
 ## v2.9.0-beta.2
 
 ### Bug fixes and improvements